Recreate caches because we rolled back a migration that was added
On https://github.com/edx/edx-platform/pull/18532 this went to master
and was picked up and added to the bok choy cache.
New work that adds a precursor migration (to remove foreign keys) fails
because oauth2_provider 0005 and 0006 can't run before we run
oauth_dispatch migrations that drop foreign keys.